pyproject.toml
[tool.poetry]
name = "online-cinema-api"
version = "0.1.0"
description = "RESTful API for an online cinema platform built with FastAPI, PostgreSQL, Celery, and Stripe"
authors = ["Your Name <your@email.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.14"
fastapi = "^0.135"
uvicorn = {extras = ["standard"], version = "^0.34"}
gunicorn = "^23.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.31"
alembic = "^1.14"
pyjwt = "^2.9"
bcrypt = "^5.0"
pydantic-settings = "^2.7"
email-validator = "^2.2"
httpx = "^0.28"
aiosmtplib = "^5.1"
aioboto3 = "^15.5"
celery = {extras = ["redis"], version = "^5.6"}
stripe = "^15.0"
jinja2 = "^3.1"
python-multipart = "^0.0.22"
sqladmin = "^0.20"
loguru = "^0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.25"
ruff = "^0.15"
mypy = "^1.13"
pre-commit = "^4.0"

[tool.ruff]
target-version = "py314"
line-length = 88
src = ["src"]
exclude = ["src/migrations/versions/"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "T20",  # flake8-print
    "RUF",  # ruff-specific rules
]
ignore = [
    "S101",  # allow assert in tests
]


[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["pydantic.mypy"]
mypy_path = "src"
explicit_package_bases = true
implicit_reexport = true
exclude = ["src/migrations/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
